gf Remove language required
